+++ provider/source/src/cldr_serde/currencies/mod.rs
@@ -3,4 +3,3 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 pub mod data;
-pub mod supplemental;
