+++ rust-toolchain.toml
@@ -3,5 +3,4 @@
 # (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 [toolchain]
-# Version updated on 2025-08-07
-channel = "1.89"
+channel = "1.90"
